Added ivar to track original tableview insets
This addresses an issue with the keyboard display/hide that interfered
with layouts that have existing content insets.

For example in a layout that is presented by a navigation controller
with `automaticallyAdjustsScrollViewInsets` enabled (the default), the
view is automatically given `contentInset` of `-64` at the `top`.

This update maintains that `-64` instead of resetting to zero when
dismissing the keyboard.